+++ src/Filter/ExistsFilter.php
@@ -4,34 +4,11 @@
 namespace Novaway\ElasticsearchClient\Filter;
 
 
-use Novaway\ElasticsearchClient\Query\CombiningFactor;
-use Webmozart\Assert\Assert;
+use Novaway\ElasticsearchClient\Query\Term\ExistsQuery;
 
-class ExistsFilter implements Filter
+/**
+ * @deprecated use Novaway\ElasticsearchClient\Query\Term\ExistsQuery instead
+ */
+class ExistsFilter extends ExistsQuery
 {
-    /** @var string */
-    protected $property;
-    /** @var string */
-    private $combiningFactor;
-
-    public function __construct(string $property, $combiningFactor = CombiningFactor::FILTER)
-    {
-        Assert::oneOf($combiningFactor, CombiningFactor::toArray());
-        $this->property = $property;
-        $this->combiningFactor = $combiningFactor;
-    }
-
-    public function getCombiningFactor(): string
-    {
-        return $this->combiningFactor;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function formatForQuery(): array
-    {
-        return ['exists' => [ 'field' => $this->property]];
-    }
-
 }
