+++ src/Filter/InArrayFilter.php
@@ -2,46 +2,11 @@
 
 namespace Novaway\ElasticsearchClient\Filter;
 
-use Novaway\ElasticsearchClient\Query\CombiningFactor;
-use Webmozart\Assert\Assert;
+use Novaway\ElasticsearchClient\Query\Term\InArrayQuery;
 
-class InArrayFilter implements Filter
+/**
+ * @deprecated use Novaway\ElasticsearchClient\Query\Term\InArrayQuery instead
+ */
+class InArrayFilter extends InArrayQuery
 {
-    /** @var string */
-    private $property;
-    /** @var array */
-    private $values;
-    /** @var string */
-    private $combiningFactor;
-
-    public function __construct(string $property, array $values, string $combiningFactor = CombiningFactor::FILTER)
-    {
-        Assert::oneOf($combiningFactor, CombiningFactor::toArray());
-        $this->property = $property;
-        $this->values = $values;
-        $this->combiningFactor = $combiningFactor;
-    }
-
-    /**
-     * @return string
-     */
-    public function getCombiningFactor(): string
-    {
-        return $this->combiningFactor;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function formatForQuery(): array
-    {
-        return [
-            'bool' => [
-                'should' => array_map(function ($value) {
-                    $matchFilter = new TermFilter($this->property, $value);
-                    return $matchFilter->formatForQuery();
-                }, $this->values)
-            ]
-        ];
-    }
 }
